requirements.txt
python-dotenv

uvicorn
fastapi
pydantic

pymongo
motor

passlib[bcrypt]